requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
alembic==1.13.2
asgiref==3.8.1
attrs==24.1.0
awscli==1.33.36
backports.tarfile==1.2.0
boto3==1.34.154
botocore==1.34.154
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.1
coreforecast==0.0.12
cycler==0.12.1
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==7.1.0
docutils==0.16
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
idna==3.7
image==1.5.33
intel-openmp==2021.4.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==8.7
keyrings.alt==5.0.1
kiwisolver==1.4.5
lightning-utilities==0.11.6
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.3
neuralforecast==1.7.4
numpy==1.26.4
opencv-python==4.10.0.84
optuna==3.6.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.3
psutil==6.0.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.3
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
ray==2.34.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.32
sqlparse==0.5.1
sympy==1.13.1
tabulate==0.9.0
tbb==2021.13.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch==2.3.1
torchmetrics==1.4.1
torchvision==0.18.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.73
ultralytics-thop==2.0.0
urllib3==2.2.2
utilsforecast==0.2.3
voluptuous==0.15.2
yarl==1.9.4
yolo==0.3.1